style: :bricks: move /shemas to app root
