--- qbt_flow_utils/config/schemas/__init__.py
@@ -1,7 +1,0 @@
-"""init schemas"""
-from qbt_flow_utils.config.schemas.clients import ClientConfig
-from qbt_flow_utils.config.schemas.scoring import ScoringConfig
-from qbt_flow_utils.config.schemas.tags import TagsConfig
-from qbt_flow_utils.config.schemas.trackers import TrackerConfig
-
-__all__ = ["ClientConfig", "ScoringConfig", "TagsConfig", "TrackerConfig"]
